pyproject.toml
[project]
name = "e-commerce-api"
version = "0.1.0"
description = "API for e-commerce applications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Django==5.2.1",
    "djangorestframework==3.16.0",
    "django-filter==25.1",
    "dotenv==0.9.9",
    "psycopg2-binary==2.9.10",
    "drf-spectacular==0.28.0",
    "django-filter==25.1",
    "djangorestframework-simplejwt>=5.5.1",
]
